fix: corrections pull requests
