+++ app/Http/Controllers/CategoriesController.php
@@ -2,16 +2,16 @@
 
 namespace App\Http\Controllers;
 
-use Illuminate\Http\Request;
-//Models
+use App\Http\Requests\Category\DeleteCategoryFormRequest;
+use App\Http\Requests\Category\ShowCategoryFormRequest;
+use App\Http\Requests\Category\StoreCategoryFormRequest;
+use App\Http\Requests\Category\UpdateCategoryFormRequest;
 use App\Models\Category;
-//FormRequest
-use App\Http\Requests\StoreCategoryFormRequest;
-use App\Http\Requests\ShowCategoryFormRequest;
-use App\Http\Requests\UpdateCategoryFormRequest;
-use App\Http\Requests\DeleteCategoryFormRequest;
-//Repositories
 use App\Repositories\CategoriesRepository;
+
+//Models
+//FormRequest
+//Repositories
 
 class CategoriesController extends Controller
 {
@@ -32,15 +32,15 @@ class CategoriesController extends Controller
     public function store(StoreCategoryFormRequest $request)
     {
         //TODO Implementar StoreCategoryFormRequest COMPLETED
-        
+
         $input = $request->validated();
-        
+
 
         $input = new Category();
         $input-> name = $request -> name;
         $input-> color = $request -> color;
         $this->categoryRepository->create($input);
-        
+
 
         return redirect()->route('categories.index')->with('success', 'Nueva Categoria Agregada!');
 
@@ -50,15 +50,15 @@ class CategoriesController extends Controller
     {
         //TODO Implementar ShowCategoryFormRequest COMPLETED
         $category = $this->categoryRepository->show($id);
-        
+
         if(!$category){
             return redirect()->route('categories.index')->with('error', 'Categoria no Encontrada');
-            
+
         }
         $validatedData = $request->validated();
         return view('categories.show', ['category' => $category]);
-        
-    
+
+
     }
 
     public function update(UpdateCategoryFormRequest $request, string $id)
@@ -78,11 +78,11 @@ class CategoriesController extends Controller
 
     public function destroy(DeleteCategoryFormRequest $request, $category)
     {
-        //TODO Implementar DeleteFormRequest COMPLETED, 
+        //TODO Implementar DeleteFormRequest COMPLETED,
 
         //validar que se elimine solo categorias que no tenga tareas asociadas PENDIENTE
         $category = $this->categoryRepository->destroy($category);
-        
+
         $category->todos()->each(function($todo){
             $todo->delete();
         });
