+++ app/Http/Controllers/UsersController.php
@@ -2,16 +2,16 @@
 
 namespace App\Http\Controllers;
 
-use Illuminate\Http\Request;
-//Models
+use App\Http\Requests\Users\DeleteUsersFormRequest;
+use App\Http\Requests\Users\ShowUsersFormRequest;
+use App\Http\Requests\Users\StoreUsersFormRequest;
+use App\Http\Requests\Users\UpdateUsersFormRequest;
 use App\Models\User;
-//FormRequest
-use App\Http\Requests\StoreUsersFormRequest;
-use App\Http\Requests\ShowUsersFormRequest;
-use App\Http\Requests\UpdateUsersFormRequest;
-use App\Http\Requests\DeleteUsersFormRequest;
-//Repositories
 use App\Repositories\UsersRepository;
+
+//Models
+//FormRequest
+//Repositories
 class UsersController extends Controller
 {
     //
@@ -22,7 +22,7 @@ class UsersController extends Controller
     public function index()
     {
         $users = $this->userRepository->all();
-        
+
         return view('users.index', ['users' => $users]);
     }
 
@@ -31,11 +31,11 @@ class UsersController extends Controller
         $user = new User();
         $user->name = $request->name_user;
         $user->email = $request->email;
-        //$user->password = bcrypt($request->password); 
-        $user->password = $request->password; 
+        //$user->password = bcrypt($request->password);
+        $user->password = $request->password;
         $this->userRepository->create($user);
         $user->save();
-    
+
         return redirect()->route('users.index')->with('success', 'Usuario agregado correctamente');
     }
 
@@ -45,7 +45,7 @@ class UsersController extends Controller
         if(!$user){
            return redirect()->route('users.index')->with('error', 'El Usuario no existe');
         }
-        
+
 
         return view('users.show', ['user' => $user, 'users' => $user]);
     }
@@ -53,12 +53,12 @@ class UsersController extends Controller
     public function update(UpdateUsersFormRequest $request,$id)
     {
         $user = $this->userRepository->update($id);
-        
+
         $user -> name = $request->name_user;
         $user -> email = $request->email;
-        $user-> password = $request->password; 
+        $user-> password = $request->password;
         //$user -> password = bcrypt($request->password);
-        
+
         $user->save();
         return redirect()->route('users.index')->with('success', 'Usuario Actualizado Correctamente');
     }
@@ -75,5 +75,5 @@ class UsersController extends Controller
 
         return redirect()->route('users.index')->with('success', 'Usuario Eliminado Correctamente');
     }
-    
+
 }
