+++ app/Http/Requests/Auth/LoginFormRequest.php
@@ -1,10 +1,10 @@
 <?php
 
-namespace App\Http\Requests;
+namespace App\Http\Requests\Auth;
 
 use Illuminate\Foundation\Http\FormRequest;
 
-class ShowTodoFormRequest extends FormRequest
+class LoginFormRequest extends FormRequest
 {
     /**
      * Determine if the user is authorized to make this request.
@@ -22,15 +22,8 @@ class ShowTodoFormRequest extends FormRequest
     public function rules(): array
     {
         return [
-            //
-            'id' => 'exists:todos,id',
-        ];
-    }
-    public function messages(): array
-    {
-        
-        return[
-            'id' => 'El ID Existe'
+            'email' => 'required|email',
+            'password' => 'required|string',
         ];
     }
 }
