+++ app/Http/Requests/Category/DeleteCategoryFormRequest.php
@@ -1,10 +1,10 @@
 <?php
 
-namespace App\Http\Requests;
+namespace App\Http\Requests\Category;
 
 use Illuminate\Foundation\Http\FormRequest;
 
-class ShowUsersFormRequest extends FormRequest
+class DeleteCategoryFormRequest extends FormRequest
 {
     /**
      * Determine if the user is authorized to make this request.
@@ -22,7 +22,8 @@ class ShowUsersFormRequest extends FormRequest
     public function rules(): array
     {
         return [
-            //
+            'id' => 'required|exists:categories,id'
         ];
     }
+
 }
