+++ app/Http/Requests/Users/ShowUsersFormRequest.php
@@ -1,10 +1,10 @@
 <?php
 
-namespace App\Http\Requests;
+namespace App\Http\Requests\Users;
 
 use Illuminate\Foundation\Http\FormRequest;
 
-class ShowCategoryFormRequest extends FormRequest
+class ShowUsersFormRequest extends FormRequest
 {
     /**
      * Determine if the user is authorized to make this request.
@@ -22,20 +22,7 @@ class ShowCategoryFormRequest extends FormRequest
     public function rules(): array
     {
         return [
-            /*
-            DUDAS
-            -Mandar error a la vista. Al no cumplir validacion
-            -Sintaxis de validacion de FormRequest
-            */
-            'id' =>  'exists:categories,id'
-
+            //
         ];
     }
-    public function messages(): array
-    {
-        return[
-            'id.required' => 'El id no existe'
-        ];
-    }
-    
 }
