+++ routes/api.php
@@ -1,12 +1,12 @@
 <?php
 
-use Illuminate\Http\Request;
-use Illuminate\Support\Facades\Route;
-use App\Http\Controllers\Api\V1\TodoController as TodoV1;
+use App\Http\Controllers\Api\Auth\LoginController;
 use App\Http\Controllers\Api\V1\CategoryController as CategoryV1;
-use App\Http\Controllers\Api\V2\TodoController as TodoV2;
+use App\Http\Controllers\Api\V1\TodoController as TodoV1;
 use App\Http\Controllers\Api\V2\CategoryController as CategoryV2;
-use App\Http\Controllers\Api\LoginController;
+use App\Http\Controllers\Api\V2\TodoController as TodoV2;
+use Illuminate\Support\Facades\Route;
+
 /*
 |--------------------------------------------------------------------------
 | API Routes
@@ -18,20 +18,23 @@ use App\Http\Controllers\Api\LoginController;
 |
 */
 //API V1
-Route::middleware('auth:sanctum')->group(function (){
-    Route::apiResource('v1/todos', Todov1::class)
-        ->only(['index','show','destroy']);
-    Route::apiResource('v1/categories',CategoryV1::class)
-        ->only(['index','show','destroy']);
+Route::prefix('v1')->group(function () {
+    Route::middleware('auth:sanctum')->group(function () {
+        Route::apiResource('todos', Todov1::class)
+            ->only(['index', 'show', 'destroy']);
+        Route::apiResource('categories', CategoryV1::class)
+            ->only(['index', 'show', 'destroy']);
+    });
 });
 //API V2
-Route::middleware('auth:sanctum')->group(function (){
-    Route::apiResource('v2/todos', TodoV2::class)
-        ->only('index', 'show', 'destroy');
-Route::apiResource('v2/categories', CategoryV2::class)
-    ->only('index', 'show', 'destroy');
+Route::prefix('v2')->group(function () {
+    Route::middleware('auth:sanctum')->group(function () {
+        Route::apiResource('todos', TodoV2::class)
+            ->only('index', 'show', 'destroy');
+        Route::apiResource('categories', CategoryV2::class)
+            ->only('index', 'show', 'destroy');
+    });
 });
-
 Route::post('/login', [LoginController::class, 'Login']);
